// frontend/src/app/interfaces/society.ts
export interface Society {
    id?: string;
    nombre: string;
    cif?: string;
    correo_electronico: string;
    tipo_sociedad: string;
    direccion?: string;
    poblacion?: string;
    pais?: string;
    codigo_postal?: string;
    codigo_sociedad?: string;
    telefono?: string;
    fax?: string;
    movil?: string;
    iban?: string;
    banco?: string;
    sucursal?: string;
    dc?: string;
    numero_cuenta?: string;
    swift?: string;
    dominio?: string;
    observaciones?: string;
    logo?: string;
    sociedad_padre_id?: string;
}